+++ test/Unit/models/classes/class.ResponseVariableTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace oat\taoResultServer\test\Unit\Models\classes;
+namespace oat\taoResultServer\test\Unit\models\classes;
 
 use PHPUnit\Framework\TestCase;
 
@@ -24,24 +24,26 @@ use PHPUnit\Framework\TestCase;
  * Copyright (c) 2020 (original work) Open Assessment Technologies S.A.
  */
 
-class taoResultServer_models_classes_TraceVariableTest extends TestCase
+class taoResultServer_models_classes_ResponseVariableTest extends TestCase
 {
     public function testVariableCanBeJsonSerialized(): void
     {
-        $subject = (new \taoResultServer_models_classes_TraceVariable())
+        $subject = (new \taoResultServer_models_classes_ResponseVariable())
             ->setIdentifier('testIdentifier')
             ->setCardinality('single')
             ->setBaseType('testBaseType')
             ->setEpoch('testEpoch')
-            ->setTrace('testTrace');
+            ->setCorrectResponse('testCorrectResponse')
+            ->setCandidateResponse('testCandidateResponse');
 
         $this->assertSame(json_encode([
             'identifier' => 'testIdentifier',
             'cardinality' => 'single',
             'baseType' => 'testBaseType',
             'epoch' => 'testEpoch',
-            'type' => \taoResultServer_models_classes_TraceVariable::TYPE,
-            'trace' => 'testTrace',
+            'type' => \taoResultServer_models_classes_ResponseVariable::TYPE,
+            'correctResponse' => 'testCorrectResponse',
+            'candidateResponse' => base64_encode('testCandidateResponse'),
         ]), json_encode($subject));
     }
 }
